+++ test/PHPUnit/connecteur/purge/PurgeTest.php
@@ -87,7 +87,6 @@ class PurgeTest extends PastellTestCase
         bool $expected_true,
         string $message
     ) {
-
         $document_info = $this->createDocument($document_type);
         $id_d = $document_info['id_d'];
 
@@ -163,11 +162,13 @@ class PurgeTest extends PastellTestCase
         $purge = $this->getObjectInstancier()->getInstance(Purge::class);
 
         $connecteurConfig = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
-        $connecteurConfig->setTabData([
+        $connecteurConfig->setTabData(
+            [
                 'actif' => 1,
                 'document_type' => 'actes-generique',
                 'document_etat' => 'creation',
-            ] + $aditionnalConnectorConfig);
+            ] + $aditionnalConnectorConfig
+        );
 
 
         $purge->setConnecteurInfo(['id_e' => 1, 'id_ce' => 42]);
@@ -416,9 +417,79 @@ class PurgeTest extends PastellTestCase
         ]);
 
         $purge->setConnecteurConfig($connecteurConfig);
-        $this->assertNotNull($this->getDonneesFormulaireFactory()->get($info_document['id_d'])->getFileContent('arrete'));
+        $this->assertNotNull(
+            $this->getDonneesFormulaireFactory()->get($info_document['id_d'])->getFileContent('arrete')
+        );
         $purge->purgerGlobal();
         DocumentSQL::clearCache();
-        $this->assertFalse($this->getObjectInstancier()->getInstance(DocumentSQL::class)->getInfo($info_document['id_d']));
+        $this->assertFalse(
+            $this->getObjectInstancier()->getInstance(DocumentSQL::class)->getInfo($info_document['id_d'])
+        );
+    }
+
+    /**
+     * @throws UnrecoverableException
+     * @throws Exception
+     */
+    public function testPurgeExclureEtat()
+    {
+        $result = $this->getInternalAPI()->post(
+            '/Document/' . PastellTestCase::ID_E_COL,
+            ['type' => 'actes-generique']
+        );
+        $id_d = $result['id_d'];
+
+        $this->getInternalAPI()->patch("/entite/1/document/$id_d", ['objet' => 'test']);
+        $this->getInternalAPI()->post(
+            "/Document/" . PastellTestCase::ID_E_COL,
+            ['type' => 'actes-generique']
+        );
+
+        $purge = $this->getObjectInstancier()->getInstance(Purge::class);
+        $connecteurConfig = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
+        $connecteurConfig->setTabData([
+            'actif' => 1,
+            'document_type' => 'actes-generique',
+            'passer_par_l_etat' => Purge::GO_TROUGH_STATE,
+            'document_etat' => 'creation',
+        ]);
+        $purge->setConnecteurInfo(['id_e' => 1, 'id_ce' => 42]);
+        $purge->setConnecteurConfig($connecteurConfig);
+        static::assertCount(2, $purge->listDocument());
+        $connecteurConfig->setTabData([
+            'document_exclure_etat' => 'modification'
+        ]);
+        static::assertCount(1, $purge->listDocument());
+    }
+
+    /**
+     * @throws UnrecoverableException
+     * @throws Exception
+     */
+    public function testPurgeExclureEtatOnly()
+    {
+        $result = $this->getInternalAPI()->post(
+            '/Document/' . PastellTestCase::ID_E_COL,
+            ['type' => 'actes-generique']
+        );
+        $id_d = $result['id_d'];
+
+        $this->getInternalAPI()->patch("/entite/1/document/$id_d", ['objet' => 'test']);
+        $this->getInternalAPI()->post(
+            "/Document/" . PastellTestCase::ID_E_COL,
+            ['type' => 'actes-generique']
+        );
+
+        $purge = $this->getObjectInstancier()->getInstance(Purge::class);
+        $connecteurConfig = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
+        $connecteurConfig->setTabData([
+            'actif' => 1,
+            'document_type' => 'actes-generique',
+            'document_exclure_etat' => 'modification',
+            'passer_par_l_etat' => Purge::GO_TROUGH_STATE,
+        ]);
+        $purge->setConnecteurInfo(['id_e' => 1, 'id_ce' => 42]);
+        $purge->setConnecteurConfig($connecteurConfig);
+        static::assertCount(0, $purge->listDocument());
     }
 }